refactor: move EncodingNormalizer from RubyGit to RubyGit::CommandLine
